修改64位操作系统使用SetWindowLong函数造成的窗体指针不完整的Bug和动作系统中时间与帧数关联错误的Bug，添加动作系统到游戏中，还缺少时间重置功能和新生成的指针未清理的问题。

+++ LLGameStudio/LLGameEngine/Game/UI/IUIProperty.h
@@ -390,6 +390,6 @@ class PropertyTotalTime :public IUIProperty
 public:
 	PropertyTotalTime() :IUIProperty(L"totalTime", L"4") { SetValue(defaultValue); }
 	wstring GetValue() { return to_wstring(value); };
-	void SetValue(wstring value) { this->value = WStringHelper::GetInt(value); };
-	int value;
+	void SetValue(wstring value) { this->value = WStringHelper::GetFloat(value); };
+	float value;
 };
